Added mouseClick (for testing) and solidnes
Mouse click detector is just an object which tracks mouse clicks and
does stuff with them. It will be used for testing actions.
Solidnes is a property of objects. If it is true, objects can be in
collisions. If its false, the ObjectManager won't check collisions with
them.

Smaller changes:
- added an empty texture to TextureContainer
- added an initializer forTextureContainer
- made small dot do its job, now with the new actions